Normative: tighten language in mergeFields()
This pulls as much logic as possible out of the calendar-dependent
operation into algorithm steps that are common to all calendars, in order
to address the implementor questions about the previous definition of the
operation.

Note, there is the potential for a rebase conflict with #2310, which
removes one use of DefaultCalendarMergeFields and renames it to
ISOCalendarMergeFields.